+++ packages/cli/src/next/builders/index.ts
@@ -1,5 +1,9 @@
 export { createBundler } from './bundler';
 export { createPhpBuilder } from './php';
+export type {
+	CreatePhpBuilderOptions,
+	PhpDriverConfigurationOptions,
+} from './php';
 export { createTsBuilder } from './ts';
 export { createPatcher } from './patcher';
 export { createPhpDriverInstaller } from '@wpkernel/php-driver';